add-guid-to-initialization
This is important for recovery purposes in cases that there are multiple triggers with guids on a single table it would be impossibly to figure out if the trigger that is initialized is the one you expected to be initialized in polling the table. Therefore, I ask that this be something that the user can pass in so that we can define the guid and track it throughout the process.